better explore page (more efficient in terms of implementation but too many new files-careful)
