+++ .gitignore
@@ -1,3 +1,6 @@
 /Dummy_Medarxiv.db
 /Dummydata.csv
 /filtered_medrxiv_papers.csv
+.vscode/settings.json
+ppc.db
+combined_db_with_ppc_id.csv
