+++ frontend/src/components/Paper/CitationChart.module.css
@@ -1,0 +1,5 @@
+.chartContainer {
+  position: relative;
+  height: 300px;
+  width: 100%;
+}
